// src/definitions/AlexaService.d.ts
    export interface AlexaRequestBody extends RequestBody {
        /** The version specifier for the request with the value defined as: “1.0” */
        version: string;
        /** The session object provides session specific information associated with the request. */
        session: Session;
        /** The context object provides data about the context of a request **/
        context: Context;
        /** An object that is composed of associated parameters that further describes the user’s request. */
        request: Request;
    }

    /**
     * Session Object
     */
    export interface Session {
        /** A boolean value indicating whether this is a new session. Returns true for a new session or false for an existing session. */
        "new": boolean;
        /** A string that represents a unique identifier per a user’s active session. Note: A sessionId is consistent for multiple subsequent requests for a user and session. If the session ends for a user, then a new unique sessionId value is provided for subsequent requests for the same user. */
        sessionId: string;
        /** A map of key-value pairs. The attributes map is empty for requests where a new session has started with the attribute new set to true.
         * The key is a string that represents the name of the attribute. Type: string
         * The value is an object that represents the value of the attribute. Type: object
         */
        attributes: any;
        /** An object containing an application ID. This is used to verify that the request was intended for your service. */
        application: {
            applicationId: string;
        };
        /**
         * An object that describes the user making the request.
         * @see User
         */
        user: User;
    }

    /**
     * Context Object
     */
    export interface Context {
        System: {
            application: {
                applicationId: string
            },
            user: {
                userId: string,
                permissions: {
                    consentToken: string
                },
                accessToken: string
            },
            device: {
                deviceId: string,
                supportedInterfaces: {
                    AudioPlayer: any
                }
            }
        };
        AudioPlayer: {
            token: string,
            offsetInMilliseconds: number,
            playerActivity: string
        };
    }

    /** User object used in Session */
    export interface User {
        /** A string that represents a unique identifier for the user who made the request. The length of this identifier can vary, but is never more than 255 characters. The userId is automatically generated when a user enables the skill in the Alexa app. Note that disabling and re-enabling a skill generates a new identifier. */
        userId: string;
        /** A token identifying the user in another system. This is only provided if the user has successfully linked their account. See [Linking an Alexa User with a User in Your System](https://developer.amazon.com/public/solutions/alexa/alexa-skills-kit/docs/linking-an-alexa-user-with-a-user-in-your-system) for more details. */
        accessToken?: string;
        /** Contains a consentToken allowing the skill access to information that the customer has consented to provide **/
        permissions?: {
            consentToken: string
        };
    }

    /**
     * Request made to an Alexa skill. This is an "abstract base class" interface.
     *
     * Called "Request" instead of just Request to show that this is not the "request" object per se, but specific to the Alexa domain.
     */
    export interface Request {
        /**
         * Type of concrete request.
         * @see AlexaRequestType
         * */
        type: AlexaRequestType;
        /** Represents the unique identifier for the specific request. */
        requestId: string;
        /** Provides the date and time when Alexa sent the request. Use this to verify that the request is current and not part of a “replay” attack. Timestamp is provided as an ISO 8601 formatted string (for example, 2015-05-13T12:34:56Z). */
        timestamp: string;
        /** A string indicating the user’s locale. For example: en-US. */
        locale: string;
    }


    /** Represents that a user made a request to an Alexa skill, but did not provide a specific intent. */
    export interface LaunchRequest extends Request {
        // Adds nothing
    }

    /** Request made to a skill based on what the user wants to do. */
    export interface IntentRequest extends Request {
        /**
         * An object that represents what the user wants.
         * @see Intent
         * */
        intent: Intent;
    }

    /** Represents an intent from a user. */
    export interface Intent {
        /** A string representing the name of the intent. */
        name: string;
        /** A map of key-value pairs that further describes what the user meant based on a predefined intent schema. The map can be empty.
         * The key is a string that describes the name of the slot. Type: string.
         * The value is an object of type slot. Type: object.
         * @see Slot */
        slots: Slot[];
    }

    /**
     * Part of intent.
     * @see Intent.
     */
    export interface Slot {
        /** A string that represents the name of the slot. */
        name: string;
        /** A string that represents the value of the slot. The value is not required.
         * Note that AMAZON.LITERAL slot values sent to your service are always in all lower case. */
        value?: string;
    }

    /** A SessionEndedRequest is an object that represents a request made to an Alexa skill to notify that a session was ended. */
    export interface SessionEndedRequest extends Request {
        /**
         * Describes why the session ended. Possible values:
         *   USER_INITIATED: The user explicitly ended the session.
         *   ERROR: An error occurred that caused the session to end.
         *   EXCEEDED_MAX_REPROMPTS: The user either did not respond or responded with an utterance that did not match any of the intents defined in your voice interface.
         * @see SessionEndedReason
         */
        reason: SessionEndedReason;
        error: {
            type: string;
            message: string;
        };
    }